feat(preserve-comments): transferring some comments from tokens to ast
Transferring comments from tokens to BaseNode, when the BaseNode is made
directly from a token. This covers a number of cases, but won't preserve all
comments. There are many places in the tree where tokens are discarded (it's an
AST after all). We will need to find these locations in the code and save the
comments. At this time we will stick with stashing comments in terminal nodes
only. Then in the formatter need to issue comment writes only where there are
known to be terminal nodes.

This checkin also uses BaseNode::default(0 in many places so so we can avoid
having to be explicit about the the fields in hundreds of test cases.