feat(preserve-comments): expanding the scope of where comments are stored
Formatting and storing comments in more terminal nodes. Also storing in Paren
Expression nodes.

Regexes need some special care. There is an unread() call once in the parser
when we've got a DIV and really want a REGEX open. At this time we need to put
the comments back as well, otherwise they are lost. To do this we stash
comments in the scanner as we move forward.

This commit also sees some field name changes. Were using commment in some
places in comments in others. Made consistent.