feat(preserve-comments): preserving comments on both sides of () in parens and calls
Using an additional vector of comments in base node to track the comments for
RHS elements in productions that have multiple non-terminals, which are not
tracked implicitly by base node.

This approach works, but has the drawback that it requires being explicit about
how many comment list nulls there are in tests cases. If we move these pointers
to the actual AST nodes we can rely on Default::default() functions.