feat(preserve-comments): adding non-terminal nodes for discarded tokens
Testing out the adding of non-terminal comment nodes for the discarded tokens
that have comments in front of them. This approach works well, but
unfortunately requires quite a bit of modification of the tests. We can't use
defaults for these nodes because the AST generally does not have optional
things. Makes for a very messy default construction. But likely will move
forward with this approach.